[Rename] 패키지 정리
Related to: #15 #22
